Create listenerMultipleMarkets
This is an JavaScript listener wrapped in an HTML tag that listens for clicks on buttons in multiple markets based upon the location of the buttons in the array.  The listener fires a dataLayer push with the appropriate information.  This function solves issues of working in multiple markets with differing languages.  It also pushes data upon function load.